fixing build issue

+++ components/m5stack_printer/m5stack_printer.h
@@ -67,9 +67,12 @@ class M5StackPrinterPrintTextAction : public Action<Ts...>, public Parented<M5St
  public:
   TEMPLATABLE_VALUE(std::string, text)
   TEMPLATABLE_VALUE(uint8_t, font_size)
-  TEMPLATABLE_VALUE(float, font_size_factor)
+  TEMPLATABLE_VALUE(double, font_size_factor)
 
-  void play(Ts... x) override { this->parent_->print_text(this->text_.value(x...), this->font_size_.value(x...)*this->font_size_factor_.value(x...)); }
+  void play(Ts... x) override {
+    this->parent_->print_text(this->text_.value(x...),
+                              this->font_size_.value(x...) * this->font_size_factor_.value(x...));
+  }
 };
 
 template<typename... Ts>
@@ -96,7 +99,6 @@ class M5StackPrinterPrintBarCodeAction : public Action<Ts...>, public Parented<M
 
   void play(Ts... x) override { this->parent_->print_barcode(this->barcode_.value(x...), this->type_.value(x...)); }
 };
-
 
 }  // namespace m5stack_printer
 }  // namespace esphome
